Add mark_read method and return number of unread messages when calling list_chats

+++ src/chats/src/queries/get_messages.rs
@@ -2,7 +2,7 @@ use ic_cdk::storage;
 use crate::domain::chat::{ChatId, Message};
 use crate::domain::chat_list::ChatList;
 
-pub fn query(chat_id: ChatId, from_index: usize) -> Option<Vec<Message>> {
+pub fn query(chat_id: ChatId, from_index: u32) -> Option<Vec<Message>> {
     let chat_list: &ChatList = storage::get();
     let me = ic_cdk::caller();
     let chat = chat_list.get(chat_id, &me)?;
